+++ Source/Relativity.Transfer.SDK.Samples.Repository/FullPathWorkflow/UploadDirectory.cs
@@ -7,11 +7,10 @@ using Relativity.Transfer.SDK.Samples.Core.Configuration;
 using Relativity.Transfer.SDK.Samples.Core.Helpers;
 using Relativity.Transfer.SDK.Samples.Core.ProgressHandler;
 using Relativity.Transfer.SDK.Samples.Core.Runner;
-using Relativity.Transfer.SDK.Samples.Repository.Common;
 
 namespace Relativity.Transfer.SDK.Samples.Repository.FullPathWorkflow;
 
-[Sample((int)SampleOrder.UploadDirectory, "Upload a directory",
+[Sample(SampleOrder.UploadDirectory, "Upload a directory",
     "The sample illustrates how to implement a directory upload to a RelativityOne file share.",
     typeof(UploadDirectory),
     TransferType.UploadDirectory)]
