+++ Source/Relativity.Transfer.SDK.Samples.Repository/FullPathWorkflow/UploadDirectoryWithCustomizedRetryPolicy.cs
@@ -10,11 +10,10 @@ using Relativity.Transfer.SDK.Samples.Core.Configuration;
 using Relativity.Transfer.SDK.Samples.Core.Helpers;
 using Relativity.Transfer.SDK.Samples.Core.ProgressHandler;
 using Relativity.Transfer.SDK.Samples.Core.Runner;
-using Relativity.Transfer.SDK.Samples.Repository.Common;
 
 namespace Relativity.Transfer.SDK.Samples.Repository.FullPathWorkflow;
 
-[Sample((int)SampleOrder.UploadDirectoryWithCustomizedRetryPolicy, "A retry policy",
+[Sample(SampleOrder.UploadDirectoryWithCustomizedRetryPolicy, "A retry policy",
     "The sample illustrates the implementation of a retry policy to enhance the resilience of a transfer.",
     typeof(UploadDirectoryWithCustomizedRetryPolicy),
     TransferType.UploadDirectory)]
