+++ Source/Relativity.Transfer.SDK.Samples.Repository/FullPathWorkflow/UploadDirectoryWithExclusionPolicy.cs
@@ -13,11 +13,10 @@ using Relativity.Transfer.SDK.Samples.Core.Configuration;
 using Relativity.Transfer.SDK.Samples.Core.Helpers;
 using Relativity.Transfer.SDK.Samples.Core.ProgressHandler;
 using Relativity.Transfer.SDK.Samples.Core.Runner;
-using Relativity.Transfer.SDK.Samples.Repository.Common;
 
 namespace Relativity.Transfer.SDK.Samples.Repository.FullPathWorkflow;
 
-[Sample((int)SampleOrder.UploadDirectoryWithExclusionPolicy, "An exclusion policy",
+[Sample(SampleOrder.UploadDirectoryWithExclusionPolicy, "An exclusion policy",
     "The sample illustrates the implementation of an exclusion policy.",
     typeof(UploadDirectoryWithExclusionPolicy),
     TransferType.UploadDirectory)]
