+++ Source/Relativity.Transfer.SDK.Samples.Repository/FullPathWorkflow/UploadItems.cs
@@ -10,11 +10,10 @@ using Relativity.Transfer.SDK.Samples.Core.Configuration;
 using Relativity.Transfer.SDK.Samples.Core.Helpers;
 using Relativity.Transfer.SDK.Samples.Core.ProgressHandler;
 using Relativity.Transfer.SDK.Samples.Core.Runner;
-using Relativity.Transfer.SDK.Samples.Repository.Common;
 
 namespace Relativity.Transfer.SDK.Samples.Repository.FullPathWorkflow;
 
-[Sample((int)SampleOrder.UploadItems, 
+[Sample(SampleOrder.UploadItems, 
     "Upload items",
     "The sample illustrates how to implement files upload to a RelativityOne file share. ", 
     typeof(UploadItems),
