+++ Source/Relativity.Transfer.SDK.Samples.Repository/FullPathWorkflow/UploadToFileSharePathBasedOnWorkspaceId.cs
@@ -16,11 +16,10 @@ using Relativity.Transfer.SDK.Samples.Core.Helpers;
 using Relativity.Transfer.SDK.Samples.Core.ProgressHandler;
 using Relativity.Transfer.SDK.Samples.Core.Runner;
 using Relativity.Transfer.SDK.Samples.Core.UI;
-using Relativity.Transfer.SDK.Samples.Repository.Common;
 
 namespace Relativity.Transfer.SDK.Samples.Repository.FullPathWorkflow;
 
-[Sample((int)SampleOrder.UploadToFileSharePathBasedOnWorkspaceId, "A transfer based on a WorkspaceId",
+[Sample(SampleOrder.UploadToFileSharePathBasedOnWorkspaceId, "A transfer based on a WorkspaceId",
     "The sample illustrates the implementation of a transfer based on a WorkspaceId.",
     typeof(UploadToFileSharePathBasedOnWorkspaceId),
     TransferType.UploadDirectoryByWorkspaceId)]
