+++ Source/Relativity.Transfer.SDK.Samples.Repository/JobBasedWorkflow/UploadDirectory.cs
@@ -9,11 +9,10 @@ using Relativity.Transfer.SDK.Samples.Core.Configuration;
 using Relativity.Transfer.SDK.Samples.Core.Helpers;
 using Relativity.Transfer.SDK.Samples.Core.ProgressHandler;
 using Relativity.Transfer.SDK.Samples.Core.Runner;
-using Relativity.Transfer.SDK.Samples.Repository.Common;
 
 namespace Relativity.Transfer.SDK.Samples.Repository.JobBasedWorkflow;
 
-[Sample((int)SampleOrder.UploadDirectoryJobBasedWorkflow, "Upload a directory (using the job based workflow)",
+[Sample(SampleOrder.UploadDirectoryJobBasedWorkflow, "Upload a directory (using the job based workflow)",
 	"The sample illustrates how to implement a directory upload (using the job based workflow) to a RelativityOne file share.",
 	typeof(UploadDirectory),
 	TransferType.UploadDirectory)]
